+++ inst/include/dplyr/DataFrameVisitors.h
@@ -23,45 +23,48 @@ namespace dplyr {
                 data(data_), 
                 visitors()
             {
-                visitor_names = data.names() ;
+                visitor_names = get_DataFrame_names(data) ;
                 nvisitors =  visitor_names.size() ;
-                for( int i=0; i<nvisitors; i++){
-                    visitors.push_back( visitor(data[i]) ) ;     
-                }
+                             
+                int index = 0 ;
+                fill_all_visitors(index, data) ;
+                
             }
             DataFrameVisitors( const Rcpp::DataFrame& data_, const Rcpp::CharacterVector& names ) : 
                 data(data_), 
                 visitors()
             {
+                
                 std::string name ;
                 int n = names.size() ;
                 nvisitors = 0 ;
                 std::vector<std::string> visitor_names_ ;
                 for( int i=0; i<n; i++){
                     name = (String)names[i] ;
-                    SEXP column ; 
-                    try {
-                        column = data[name] ;
-                    } catch(...){
+                    SEXP column = get_column(data, name) ; 
+                    
+                    if( column == R_NilValue ){
                         std::stringstream s ;
                         s << "unknown column '"
                           << name
                           << "'"; 
                         stop(s.str()); 
                     }
-                    if( column != R_NilValue ){
-                        nvisitors++ ;
-                        visitor_names_.push_back( name ) ;
-                        visitors.push_back( visitor( column ) ) ;
-                    }
+                    
+                    nvisitors++ ;
+                    visitor_names_.push_back( name ) ;
+                    visitors.push_back( visitor( column ) ) ;
+                    
                 }
                 visitor_names = wrap( visitor_names_ );
+                
             } 
             
             ~DataFrameVisitors(){
                 delete_all( visitors );
-            }
-        
+            }     
+            
+            
             template <typename Container>
             DataFrame subset( const Container& index, const CharacterVector& classes ) const {
                 List out(nvisitors);
@@ -69,6 +72,7 @@ namespace dplyr {
                    out[k] = get(k)->subset(index) ;    
                 }
                 structure( out, Rf_length(out[0]) , classes) ;
+                
                 return (SEXP)out ;
             }
             
@@ -81,7 +85,23 @@ namespace dplyr {
             
         private:
             
-            void structure( List& x, int nrows, CharacterVector classes ) const {
+            inline SEXP get_column( const Rcpp::DataFrame& df, std::string& name ){
+                int n = df.size() ;
+                CharacterVector names = df.names() ;
+                for(int i=0; i<n; i++){
+                    SEXP column  = df[i] ;
+                    if( Rf_inherits( column, "data.frame" ) ){
+                        SEXP res = get_column( column, name ) ;
+                        if( res != R_NilValue ) return res ;
+                    } else {
+                        String column_name = names[i] ;
+                        if( column_name == name ) return column ;
+                    }
+                }
+                return R_NilValue ;
+            }
+            
+            inline void structure( List& x, int nrows, CharacterVector classes ) const {
                 x.attr( "class" ) = classes ;
                 set_rownames(x, nrows) ;
                 x.names() = visitor_names ;
@@ -90,10 +110,70 @@ namespace dplyr {
                     x.attr( "vars" ) = vars ;
             }
             
+            inline void fill_all_visitors( int& index, const Rcpp::DataFrame& df){
+                int n = df.size() ;
+                CharacterVector names = df.names() ;
+                for(int i=0; i<n; i++){
+                    SEXP column = df[i] ;
+                    if( Rf_inherits( column, "data.frame" ) ){
+                        fill_all_visitors( index, column ) ;        
+                    } else {
+                        visitors.push_back( visitor(column) ) ;
+                        index++ ;
+                    }
+                }
+            }
+        
+            inline int get_DataFrame_column_count( const Rcpp::DataFrame& df ){
+                int nc = 0 ;
+                
+                int n = df.size() ;
+                for( int i=0; i<n; i++){
+                    SEXP column = df[i] ;
+                    if( Rf_inherits( column, "data.frame" ) ){
+                        nc += get_DataFrame_column_count( column ) ;    
+                    } else {
+                        nc++ ;    
+                    }
+                }
+                return nc ;
+            }
+            
+            inline void fill_names( CharacterVector& out, int& index, const Rcpp::DataFrame& df){
+                int n = df.size() ;
+                CharacterVector names = df.names() ;
+                for(int i=0; i<n; i++){
+                    SEXP column = df[i] ;
+                    if( Rf_inherits( column, "data.frame" ) ){
+                        fill_names( out, index, column ) ;
+                    } else {
+                        out[index] = names[i] ;
+                        index++ ;
+                    }
+                }
+                
+            }
+            
+            inline CharacterVector get_DataFrame_names(const Rcpp::DataFrame& df ){
+                // first count how many columns
+                int nc = get_DataFrame_column_count(df) ;
+                
+                CharacterVector out(nc) ;
+                int index = 0 ;
+                fill_names(out, index, df ) ;
+                return out ;
+            }
+    
+            
     } ;
           
     inline DataFrame subset( DataFrame data, LogicalVector test, CharacterVector select, CharacterVector classes ){
         DataFrameVisitors visitors( data, select ) ;
+        return visitors.subset(test, classes ) ;
+    }
+
+    inline DataFrame subset( DataFrame data, LogicalVector test, CharacterVector classes ){
+        DataFrameVisitors visitors( data ) ;
         return visitors.subset(test, classes ) ;
     }