+++ inst/include/dplyr/white_list.h
@@ -24,7 +24,17 @@ inline bool white_list(SEXP x){
     case STRSXP:   return Rf_inherits(x, "AsIs") || is_bare_vector( x ) ;
     case CPLXSXP:  return Rf_inherits(x, "AsIs") || is_bare_vector( x ) ;
     
-    case VECSXP:   return ! Rf_inherits(x, "POSIXlt") && is_bare_vector( x ) ;
+    case VECSXP:   {
+            if( Rf_inherits(x, "data.frame" ) ){
+                // check that all variables are white listed
+                DataFrame df = x ;
+                for( int i=0; i<df.size() ; i++){
+                    if( !white_list( df[i] ) ) return false ;    
+                }
+                return true ;
+            }
+            return ! Rf_inherits(x, "POSIXlt") && is_bare_vector( x ) ;
+    }
     
     default: break ;
     }